Completed the JSX
With this commit i have created  a sidebar  with search function.
Imported assets folder for the icons applied .
Try and catch for the search bar  and search bar header was added.
Created a new jsx ChannelSearch for the logic of the search bar
Updated the index.js file with new imports.